+++ eeschema/dialogs/dialog_schematic_setup.cpp
@@ -67,7 +67,7 @@ DIALOG_SCHEMATIC_SETUP::DIALOG_SCHEMATIC_SETUP( SCH_EDIT_FRAME* aFrame ) :
 
     FinishDialogSettings();
 
-    for( int i = 0; i < m_treebook->GetPageCount(); ++i )
+    for( size_t i = 0; i < m_treebook->GetPageCount(); ++i )
    	    m_macHack.push_back( true );
 }
